content/html/content/test/file_iframe_sandbox_c_if7.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 341604</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<script type="text/javascript">
  function ok(result, desc) {
    window.parent.postMessage({ok: result, desc: desc}, "*");
  }

  function doStuff() {
    try {
      var thing = indexedDB.open("sandbox");
      ok(false, "documents sandboxed without allow-same-origin should NOT be able to access indexedDB");
    }

    catch(e) {
      ok(true, "documents sandboxed without allow-same-origin should NOT be able to access indexedDB");
    }  
  }
</script>
<body onLoad='doStuff();'>
  I am sandboxed but with "allow-scripts"
</body>
</html>
